// src/main/resources/angular/src/app/api/model/profile.ts
/**
 * MAAS API
 * This is the Mobility as a Service API
 *
 * OpenAPI spec version: 1.0.0
 * Contact: avalls@nexusgeographics.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Funcionality } from './funcionality';


export interface Profile {
    profileId: number;
    description: string;
    funcionalities: Array<Funcionality>;
}